perbaikan migration

+++ be-kinerjaprodi/database/migrations/2022_06_28_093239_create_penelitians_table.php
@@ -19,11 +19,11 @@ class CreatePenelitiansTable extends Migration
             $table->string('judul');
             $table->string('tahun');
             $table->string('sumber_dana_PT_mandiri');
-            $table->integer('dana_PT_Mandiri');
+            $table->bigInteger('dana_PT_Mandiri');
             $table->string('sumber_dalam_negri');
-            $table->integer('dana_dalam_negri');
+            $table->bigInteger('dana_dalam_negri');
             $table->string('sumber_luar_negri');
-            $table->integer('dana_luar_negri');
+            $table->bigInteger('dana_luar_negri');
             $table->timestamps();
         });
     }
